// auth-service/src/modules/admin/admin-company/dto/company.dto.ts
export class CompanyDto {
  companyCode?: string;
  taxCode?: string;
  companyName?: string;
  address?: string;
  country?: string;
  companyType?: string;
  mainIndustry?: string;
  representativeName?: string;
  startDate?: Date;
  phoneNumber?: string;
  email?: string;
  websiteAddress?: string;
  logoUrl?: string;
  status?: string;
}